Here we go...
Did not like the numeric value retrieval code. Removed the length parameters. If you need numeric data, and you are not sure if it exists, use the new HashTableHashEntry() API.

HashTableGetLastError will return now the proper error code if the table "didn't" or "does not" exist.

Linked up the prototypes from the header file, so as to keep from having to maintain two tables of enums. Also, we are making sure the prototypes match up with what we said they do in this manner.

It's likely I changed more things in this file... But my IDE is too stupid to let me scroll through my edits and remark the new changes. Modal dialogs are EVIL. They should only be used when the parent window is waiting for updates from the child window. Keep that in mind folks. Don't be a dummy, extend your functionality.